+++ gather_test.go
@@ -14,47 +14,28 @@ import (
 
 func TestListenUDP(t *testing.T) {
 	a, err := NewAgent(&AgentConfig{})
-	if err != nil {
-		t.Fatalf("Failed to create agent: %s", err)
-	}
+	assert.NoError(t, err)
 
 	localIPs, err := a.localInterfaces([]NetworkType{NetworkTypeUDP4})
-	if len(localIPs) == 0 {
-		t.Fatal("localInterfaces found no interfaces, unable to test")
-	} else if err != nil {
-		t.Fatal(err)
-	}
+	assert.NotEqual(t, len(localIPs), 0, "localInterfaces found no interfaces, unable to test")
+	assert.NoError(t, err)
 
 	ip := localIPs[0]
 
 	conn, err := a.listenUDP(0, 0, udp, &net.UDPAddr{IP: ip, Port: 0})
-	if err != nil {
-		t.Fatalf("listenUDP error with no port restriction %v", err)
-	} else if conn == nil {
-		t.Fatalf("listenUDP error with no port restriction return a nil conn")
-	}
+	assert.NoError(t, err, "listenUDP error with no port restriction")
+	assert.NotNil(t, conn, "listenUDP error with no port restriction return a nil conn")
 
 	_, err = a.listenUDP(4999, 5000, udp, &net.UDPAddr{IP: ip, Port: 0})
-	if err == nil {
-		t.Fatal("listenUDP with invalid port range did not fail")
-	}
-	if err != ErrPort {
-		t.Fatal("listenUDP with invalid port range did not return ErrPort")
-	}
+	assert.Equal(t, err, ErrPort, "listenUDP with invalid port range did not return ErrPort")
 
 	conn, err = a.listenUDP(5000, 5000, udp, &net.UDPAddr{IP: ip, Port: 0})
-	if err != nil {
-		t.Fatalf("listenUDP error with no port restriction %v", err)
-	} else if conn == nil {
-		t.Fatalf("listenUDP error with no port restriction return a nil conn")
-	}
+	assert.NoError(t, err, "listenUDP error with no port restriction")
+	assert.NotNil(t, conn, "listenUDP error with no port restriction return a nil conn")
 
 	_, port, err := net.SplitHostPort(conn.LocalAddr().String())
-	if err != nil {
-		t.Fatal(err)
-	} else if port != "5000" {
-		t.Fatalf("listenUDP with port restriction of 5000 listened on incorrect port (%s)", port)
-	}
+	assert.NoError(t, err)
+	assert.Equal(t, port, "5000", "listenUDP with port restriction of 5000 listened on incorrect port")
 
 	portMin := 5100
 	portMax := 5109
@@ -63,11 +44,8 @@ func TestListenUDP(t *testing.T) {
 	portRange := make([]int, 0, total)
 	for i := 0; i < total; i++ {
 		conn, err = a.listenUDP(portMax, portMin, udp, &net.UDPAddr{IP: ip, Port: 0})
-		if err != nil {
-			t.Fatalf("listenUDP error with no port restriction %v", err)
-		} else if conn == nil {
-			t.Fatalf("listenUDP error with no port restriction return a nil conn")
-		}
+		assert.NoError(t, err, "listenUDP error with no port restriction")
+		assert.NotNil(t, conn, "listenUDP error with no port restriction return a nil conn")
 
 		_, port, err = net.SplitHostPort(conn.LocalAddr().String())
 		if err != nil {
@@ -88,12 +66,7 @@ func TestListenUDP(t *testing.T) {
 		t.Fatalf("listenUDP with port restriction [%d, %d], got:%v, want:%v", portMin, portMax, result, portRange)
 	}
 	_, err = a.listenUDP(portMax, portMin, udp, &net.UDPAddr{IP: ip, Port: 0})
-	if err == nil {
-		t.Fatalf("listenUDP with port restriction [%d, %d], should return error", portMin, portMax)
-	}
-	if err != ErrPort {
-		t.Fatalf("listenUDP with port restriction [%d, %d], did not return ErrPort", portMin, portMax)
-	}
+	assert.Equal(t, err, ErrPort, "listenUDP with port restriction [%d, %d], did not return ErrPort", portMin, portMax)
 
 	assert.NoError(t, a.Close())
 }
